refactor: Split measurement integration tests by protocol
Split InProcessLifeOfAMeasurementIntegrationTest and
InProcessEdpAggregatorLifeOfAMeasurementIntegrationTest into
protocol-specific subclasses (Direct, HMSS, TrusTee) to isolate
flaky tests and enable independent CI reruns.
