+++ src/test/kotlin/org/wfanet/measurement/integration/deploy/gcloud/GCloudEdpAggregatorHmssMeasurementTest.kt
@@ -1,4 +1,4 @@
-// Copyright 2025 The Cross-Media Measurement Authors
+// Copyright 2026 The Cross-Media Measurement Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,24 +19,15 @@ import org.junit.Rule
 import org.junit.rules.Timeout
 import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorRule
 import org.wfanet.measurement.integration.common.ALL_DUCHY_NAMES
-import org.wfanet.measurement.integration.common.InProcessEdpAggregatorLifeOfAMeasurementIntegrationTest
+import org.wfanet.measurement.integration.common.InProcessEdpAggregatorHmssMeasurementTest
 
-/**
- * Implementation of [InProcessEdpAggregatorLifeOfAMeasurementIntegrationTest] for GCloud backends
- * with Spanner database.
- */
-class GCloudEdpAggregatorLifeOfAMeasurementIntegrationTest :
-  InProcessEdpAggregatorLifeOfAMeasurementIntegrationTest(
+class GCloudEdpAggregatorHmssMeasurementTest :
+  InProcessEdpAggregatorHmssMeasurementTest(
     kingdomDataServicesRule = KingdomDataServicesProviderRule(spannerEmulator),
     duchyDependenciesRule = SpannerDuchyDependencyProviderRule(spannerEmulator, ALL_DUCHY_NAMES),
     secureComputationDatabaseAdmin = spannerEmulator,
   ) {
 
-  /**
-   * Rule to enforce test method timeout.
-   *
-   * TODO(Kotlin/kotlinx.coroutines#3865): Switch back to CoroutinesTimeout when fixed.
-   */
   @get:Rule val timeout: Timeout = Timeout.seconds(180)
 
   companion object {
